+++ experimental/packages/sdk-logs/src/types.ts
@@ -62,15 +62,17 @@ export type LoggerConfigurator = (
   loggerScope: InstrumentationScope
 ) => Required<LoggerConfig>;
 
+export interface ForceFlushOptions {
+  /**
+   * How long the force flush can run before it is cancelled.
+   * The default value is 30000ms.
+   */
+  timeoutMillis?: number;
+}
+
 export interface LoggerProviderOptions {
   /** Resource associated with trace telemetry  */
   resource?: Resource;
-
-  /**
-   * How long the forceFlush can run before it is cancelled.
-   * The default value is 30000ms
-   */
-  forceFlushTimeoutMillis?: number;
 
   /** Log Record Limits*/
   logRecordLimits?: LogRecordLimits;
@@ -106,11 +108,24 @@ export interface LogRecordLimits {
   attributeCountLimit?: number;
 }
 
-export interface SimpleLogRecordProcessorOptions {
+/**
+ * Common options for SDK log processors.
+ */
+export interface LogRecordProcessorOptions {
+  /**
+   * A meter provider to which to record self-observability log processor metrics.
+   * @experimental This option is experimental and is subject to breaking changes in minor releases.
+   */
+  selfObsMeterProvider?: MeterProvider;
+}
+
+export interface SimpleLogRecordProcessorOptions
+  extends LogRecordProcessorOptions {
   exporter: LogRecordExporter;
 }
 
-export interface BatchLogRecordProcessorOptions {
+export interface BatchLogRecordProcessorOptions
+  extends LogRecordProcessorOptions {
   exporter: LogRecordExporter;
 
   /** The maximum batch size of every export. It must be smaller or equal to